A customizable fetch request for each separate point cloud
