+++ src/loading/index.ts
@@ -1,4 +1,3 @@
 export * from './binary-loader';
 export * from './load-poc';
 export * from './types';
-export * from './xhr-request';
