+++ src/loading/load-poc.ts
@@ -11,7 +11,6 @@ import { getIndexFromName } from '../utils/utils';
 import { Version } from '../version';
 import { BinaryLoader } from './binary-loader';
 import { GetUrlFn } from './types';
-import { XhrRequest } from './xhr-request';
 
 interface BoundingBoxData {
   lx: number;
@@ -44,18 +43,20 @@ interface POCJson {
  *    Function which receives the relative URL of a point cloud chunk file which is to be loaded
  *    and shoud return a new url (e.g. signed) in the form of a string or a promise.
  *
+ * @param xhrRequest A request for each point cloud separate request.
+ *
  * @returns
  *    An observable which emits once when the first LOD of the point cloud is loaded.
  */
-export function loadPOC(url: string, getUrl: GetUrlFn): Promise<PointCloudOctreeGeometry> {
+export function loadPOC(url: string, getUrl: GetUrlFn, xhrRequest = fetch): Promise<PointCloudOctreeGeometry> {
   return Promise.resolve(getUrl(url)).then(transformedUrl => {
-    return XhrRequest.fetch(transformedUrl)
+    return xhrRequest(transformedUrl, { mode: 'cors' })
         .then(res => res.json())
-        .then(parse(transformedUrl, getUrl));
+        .then(parse(transformedUrl, getUrl, xhrRequest));
   });
 }
 
-function parse(url: string, getUrl: GetUrlFn) {
+function parse(url: string, getUrl: GetUrlFn, xhrRequest = fetch) {
   return (data: POCJson): Promise<PointCloudOctreeGeometry> => {
     const { offset, boundingBox, tightBoundingBox } = getBoundingBoxes(data);
 
@@ -64,9 +65,10 @@ function parse(url: string, getUrl: GetUrlFn) {
       version: data.version,
       boundingBox,
       scale: data.scale,
+      xhrRequest
     });
 
-    const pco = new PointCloudOctreeGeometry(loader, boundingBox, tightBoundingBox, offset);
+    const pco = new PointCloudOctreeGeometry(loader, boundingBox, tightBoundingBox, offset, xhrRequest);
 
     pco.octreeDir = data.octreeDir.indexOf('http') === 0 ? data.octreeDir : `${url}/../${data.octreeDir}`;
     pco.url = url;
