+++ src/loading/types.ts
@@ -1,1 +1,2 @@
 export type GetUrlFn = (url: string) => string | Promise<string>;
+export type XhrRequest = (input: RequestInfo, init?: RequestInit) => Promise<Response>;
